requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
moviepy==1.0.3
yt-dlp==2023.12.30
requests==2.31.0
openai-whisper==20231117
torch==2.1.2
numpy==1.26.2
